requirements.txt
pymongo[srv]
python-dotenv
fastapi
uvicorn
requests
python-telegram-bot
langchain-openai
motor # for mongodb async
